sources/vlm_parser/requirements.txt
# Core dependencies for VLM PDF parser
python-dotenv>=1.0.0
openai>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
requests>=2.31.0
PyMuPDF>=1.24.7
boto3>=1.34.0

# Additional document-related dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Business plan creation
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
strands-agents-builder

python-docx>=0.8.11
openpyxl>=3.1.0